fix: Address deprecated code and missing imports
- GoLLIE/src/model/load_model.py: Change 'dtype' to 'torch_dtype' parameter
  (deprecated in transformers, causes warnings)

- GoLLIE/src/trainer.py: Fix incorrect logging.warning() usage for deprecation
  warning. The DeprecationWarning class and stacklevel should be passed to
  warnings.warn(), not logging.warning()

- run_gollie_experiments.py: Add missing imports for logging, re, inspect,
  and black that are used but not imported

+++ GoLLIE/src/model/load_model.py
@@ -390,7 +390,7 @@ def load_model(
         device_map=device_map,
         max_memory=max_memory,
         quantization_config=bnb_config,
-        dtype=torch_dtype,
+        torch_dtype=torch_dtype,
         config=config,
         trust_remote_code=trust_remote_code,
     )
